Update lstm.cpp
Added a function that transforms our list of notes into a list of lists of notes (currently set to 4 but is trivial to add parameter that allows us to change). Disclaimer: I am still to update the training function to take into consideration the new shape of the data. I will do this in my next commit. (Creating a new branch as I'm not sure if this is relevant)